+++ .devcontainer/post_create.sh
@@ -1,1 +1,2 @@
 git config --global --add safe.directory /root/tj2_ros
+git config --global --add safe.directory /root/tj2_ros/docs/tj2_ros.wiki
